Make sure response changes are persistent.
